requirements.txt
pyyaml
jinja2
mkdocs
mkdocs-material
